+++ src/components/nav-drawer/themes/header-item.ts
@@ -3,11 +3,19 @@ import type { Themes } from '../../../theming/types.js';
 
 // Shared Styles
 import { styles as bootstrap } from './shared/header-item/header-item.bootstrap.css.js';
+import { styles as fluent } from './shared/header-item/header-item.fluent.css.js';
 import { styles as indigo } from './shared/header-item/header-item.indigo.css.js';
+import { styles as material } from './shared/header-item/header-item.material.css.js';
 
 const light = {
   bootstrap: css`
     ${bootstrap}
+  `,
+  material: css`
+    ${material}
+  `,
+  fluent: css`
+    ${fluent}
   `,
   indigo: css`
     ${indigo}
@@ -17,6 +25,12 @@ const light = {
 const dark = {
   bootstrap: css`
     ${bootstrap}
+  `,
+  material: css`
+    ${material}
+  `,
+  fluent: css`
+    ${fluent}
   `,
   indigo: css`
     ${indigo}